Moved RobotClasses used by Robot.py into dir
